add minimap2 to QC and BioReaders handle minimap2 sam

+++ SequelQC/SIRVvalidate_smrtlink_isoseq.py
@@ -12,31 +12,58 @@ GMAP_BIN = "/home/UNIXHOME/etseng/bin/gmap"
 GMAP_DB = "/home/UNIXHOME/etseng/share/gmap_db_new/"
 GMAP_CPUS = 12
 SIRV_DIR = "/home/UNIXHOME/etseng/projects2016/Lexogen_SIRV/ground_truth/"
+SIRV_GENOME = "/home/UNIXHOME/etseng/projects2015/Lexogen_SIRV/ground_truth/SIRV_150601a.fasta"
 
 STAR_BIN = "python /home/UNIXHOME/etseng/GitHub/cDNA_Cupcake/sequence/STARwrapper.py"
 SIRV_STAR_DB = "/home/UNIXHOME/etseng/share/star_db/SIRV_with_annotation/"
 
+HG38_GENOME = "/pbi/dept/bifx/etseng/genomes/hg38/hg38.fa"
+
 def link_files(src_dir, out_dir):
     os.makedirs(out_dir)
+    # location for HQ fastq in IsoSeq1
     hq_fastq = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript.tasks.combine_cluster_bins-0', 'hq_isoforms.fastq')
+    # location for HQ fastq in IsoSeq2
+    hq_fastq2 = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript2tools.tasks.collect_polish-0', 'all_arrowed_hq.fastq')
+    # location for cluster report in IsoSeq1
     cluster_csv = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript.tasks.combine_cluster_bins-0', 'cluster_report.csv')
+    cluster_csv2 = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript2tools.tasks.collect_polish-0', 'report.csv')
 
-    os.symlink(hq_fastq, os.path.join(out_dir, 'hq_isoforms.fastq'))
-    os.symlink(cluster_csv, os.path.join(out_dir, 'cluster_report.csv'))
+    if os.path.exists(hq_fastq):
+        print >> sys.stderr, "Detecting IsoSeq1 task directories..."
+        os.symlink(hq_fastq, os.path.join(out_dir, 'hq_isoforms.fastq'))
+        os.symlink(cluster_csv, os.path.join(out_dir, 'cluster_report.csv'))
+        isoseq_version = '1'
+    else:
+        print >> sys.stderr, "Detecting IsoSeq2 task directories..."
+        os.symlink(hq_fastq2, os.path.join(out_dir, 'hq_isoforms.fastq'))
+        os.symlink(cluster_csv2, os.path.join(out_dir, 'cluster_report.csv'))
+        isoseq_version = '2'
+    return out_dir, 'hq_isoforms.fastq', 'cluster_report.csv', isoseq_version
 
-    return out_dir, 'hq_isoforms.fastq', 'cluster_report.csv'
-
-def make_abundance_from_Sequel_cluster_csv(cluster_csv, collapse_prefix):
+def make_abundance_from_Sequel_cluster_csv(cluster_csv, collapse_prefix, isoseq_version):
     """
+    in Iso-Seq1:
     cluster_id,read_id,read_type
     i0_ICE_samplee5686f|c1,m54011_160718_221804/53150541/967_60_CCS,FL
+  
+    in IsoSeq2:
+    cluster_id,read_id,read_type
+    cb10060_c58,m54086_170204_081430/41025830/1758_53_CCS,FL
+    
+    (hq id example: HQ_polishOFF|cb5925_c32246/f3p0/1332)
     """
     fl_ass = defaultdict(lambda: set()) # (i0,c13) -> # of FL associated
     reader = DictReader(open(cluster_csv),delimiter=',')
     for r in reader:
-        a,b=r['cluster_id'].split('|')
-        cid=b
-        pre=a.split('_')[0]
+        if isoseq_version=='1':
+            a,b=r['cluster_id'].split('|')
+            cid=b
+            pre=a.split('_')[0]
+        elif isoseq_version=='2':
+            pre,cid = r['cluster_id'].split('_')
+        else:
+            raise Exception, "Unrecorgnized isoseq version {0}!".format(isoseq_version)
         # current version of SMRTLink does not provide nFL information =____=
         if r['read_type']=='FL':
             fl_ass[(pre,cid)].add(r['read_id'])
@@ -49,9 +76,13 @@ def make_abundance_from_Sequel_cluster_csv(cluster_csv, collapse_prefix):
         pbid,members=line.strip().split('\t')
         total = 0
         for m in members.split(','):
-            raw = m.split('|')
-            cid = raw[1].split('/')[0]
-            pre = raw[0].split('_')[0]
+            if isoseq_version=='1':
+                raw = m.split('|')
+                cid = raw[1].split('/')[0]
+                pre = raw[0].split('_')[0]
+            elif isoseq_version=='2':
+                # ex: HQ_polishOFF|cb5925_c32246/f3p0/1332
+                pre, cid = m.split('|')[1].split('/')[0].split('_')
             total += len(fl_ass[(pre,cid)])
         f.write("{0}\t{1}\n".format(pbid, total))
     f.close()
@@ -61,7 +92,7 @@ def sanity_check_script_dependencies():
         print >> sys.stderr, "chain_samples.py required in PATH! Please install Cupcake ToFU!"
         sys.exit(-1)
 
-def collapse_to_SIRV(out_dir, hq_fastq, cluster_csv, min_count, aligner_choice):
+def collapse_to_SIRV(out_dir, hq_fastq, cluster_csv, min_count, aligner_choice, isoseq_version):
 
     cur_dir = os.getcwd()
     os.chdir(out_dir)
@@ -72,6 +103,9 @@ def collapse_to_SIRV(out_dir, hq_fastq, cluster_csv, min_count, aligner_choice):
     elif aligner_choice=='star':
         cmd = "{star} {db} {hq} {hq}.sam --cpus {cpus}".format(\
             star=STAR_BIN, db=SIRV_STAR_DB, hq=hq_fastq, cpus=GMAP_CPUS)
+    elif aligner_choice=='minimap2':
+        cmd = "minimap2 -ax splice -t {cpus} {ref} {hq} > {hq}.sam 2> {hq}.sam.log".format(\
+            cpus=GMAP_CPUS, ref=SIRV_GENOME, hq=hq_fastq) 
 
     if subprocess.check_call(cmd, shell=True)!=0:
         raise Exception, "ERROR CMD:", cmd
@@ -87,7 +121,7 @@ def collapse_to_SIRV(out_dir, hq_fastq, cluster_csv, min_count, aligner_choice):
 
     ### make_abundance_from_CSV
     collapse_prefix = hq_fastq + '.no5merge.collapsed'
-    make_abundance_from_Sequel_cluster_csv(cluster_csv, collapse_prefix)
+    make_abundance_from_Sequel_cluster_csv(cluster_csv, collapse_prefix, isoseq_version)
 
 
 
@@ -165,12 +199,12 @@ if __name__ == "__main__":
     parser.add_argument("--tmp_dir", default="tmp", help="tmp dirname (default: tmp)")
     parser.add_argument("--eval_dir", default="eval", help="eval dirname (default: eval)")
     parser.add_argument("--min_count", type=int, default=2, help="min FL count cutoff (default:2)")
-    parser.add_argument("--aligner_choice", default='star', choices=('gmap', 'star'), help="Aligner choice (default: star)")
+    parser.add_argument("--aligner_choice", default='star', choices=('gmap', 'star', 'minimap2'), help="Aligner choice (default: star)")
 
     args = parser.parse_args()
     sanity_check_script_dependencies()
 
-    o, a, b = link_files(args.smrtlink_dir, args.tmp_dir)
-    collapse_to_SIRV(o, a, b, args.min_count, args.aligner_choice)
+    o, a, b, v = link_files(args.smrtlink_dir, args.tmp_dir)
+    collapse_to_SIRV(o, a, b, args.min_count, args.aligner_choice, v)
     validate_with_SIRV(args.tmp_dir, args.eval_dir)
     eval_result(args.eval_dir, args.smrtlink_dir, args.min_count)
